comit2

+++ database/migrations/2024_05_07_011639_create_producto_table.php
@@ -1,5 +1,6 @@
 <?php
-
+//AGREGAR SOFT DELETE MOVIENDO EL MIGRATE PARA QUE SE AGREGUE LA COLUMNA NECESARIA 
+//SQLSTATE[42S22]: Column not found: 1054 Unknown column 'producto.deleted_at' in 'where clause'
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -15,8 +16,9 @@ return new class extends Migration
             $table->id();
             $table->string('nombre');
             $table->string('descripcion');
-            $table ->int('precio');
+            $table ->float('precio');
             $table->timestamps();
+        //    $table->softDeletes(); 
         });
     }
